requirements.txt
textX
owlready2
torch
opencv-python
clip
pillow
torchvision
tqdm
ftfy
regex
pyscipopt
matplotlib
captum
IPython